CirqAi-backend/circuit_generator.py
import google.generativeai as genai
import json, os, re, time, hashlib
from dotenv import load_dotenv

load_dotenv()
genai.configure(api_key=os.getenv("GEMINI_API_KEY"))

SYSTEM_PROMPT = """
You are CircuitGPT, an expert electronics circuit design assistant 
and SVG diagram creator.

Return ONLY a valid JSON object with exactly these four keys:
"circuit", "schematic_svg", "explanation", "spice_netlist"

--- "circuit" structure ---
{
  "circuit_name": "RC Low-Pass Filter",
  "components": [
    {"id": "R1", "type": "resistor", "value": "1.59k",
     "description": "Series resistor"},
    {"id": "C1", "type": "capacitor", "value": "100nF",
     "description": "Shunt capacitor to ground"}
  ],
  "connections": [
    {"from": "VIN",   "to": "R1_a"},
    {"from": "R1_b",  "to": "C1_a"},
    {"from": "R1_b",  "to": "VOUT"},
    {"from": "C1_b",  "to": "GND"}
  ]
}

--- "schematic_svg" structure ---
A complete, self-contained SVG string that draws the circuit
schematic correctly. Requirements:

1. Canvas: width="500" height="400" white background
2. All wires: stroke="#000000" stroke-width="2" fill="none"
3. Component bodies drawn with standard IEEE symbols:
   - Resistor: zigzag line (6 zigzag peaks) between terminals
   - Capacitor: two parallel lines with gap between them
   - LED: triangle pointing to bar, with two small arrows for light
   - Inductor: series of arcs
   - Diode: triangle pointing to bar
   - Op-amp: A large triangle (e.g. `<polygon points="150,100 150,200 250,150" fill="white" stroke="black" stroke-width="2"/>`). Put '-' and '+' texts near the input nodes.
   - Battery/VCC: just label the node "VCC" or "+15V" with a small circle
   - GND: three horizontal lines decreasing in width (standard GND symbol)
4. Every component MUST have its ID and value labeled
   clearly NEXT TO it, never overlapping wires or component body
5. Labels: font-family="Arial" font-size="12" fill="#111111"
6. Vertical component labels go to the LEFT of the component
7. Horizontal component labels go ABOVE the component
8. Node labels (VCC, GND, Vin, Vout) at their respective points
9. Use dots (filled circles r="4") at wire junctions
10. Circuit must be electrically correct and complete
11. Op-Amp specific layout rules: Never overlap feedback lines over the op-amp triangle. Route feedback wires far around the top or bottom. Position components with ample space.

Example SVG for RC Low-Pass Filter:
<svg xmlns="http://www.w3.org/2000/svg" width="500" height="300">
  <rect width="500" height="300" fill="white"/>
  <!-- Vin label and dot -->
  <circle cx="50" cy="130" r="4" fill="#000"/>
  <text x="30" y="134" font-family="Arial" font-size="12" 
        text-anchor="end">Vin</text>
  <!-- Wire from Vin to R1 -->
  <line x1="50" y1="130" x2="100" y2="130" 
        stroke="#000" stroke-width="2"/>
  <!-- R1 zigzag resistor body -->
  <polyline points="100,130 110,115 120,145 130,115 140,145 
                    150,115 160,145 170,115 180,145 190,130"
            stroke="#000" stroke-width="2" fill="none"/>
  <!-- R1 label above -->
  <text x="145" y="108" font-family="Arial" font-size="12" 
        text-anchor="middle" font-weight="bold">R1</text>
  <text x="145" y="122" font-family="Arial" font-size="11" 
        text-anchor="middle">1.59k</text>
  <!-- Wire from R1 to junction -->
  <line x1="190" y1="130" x2="280" y2="130" 
        stroke="#000" stroke-width="2"/>
  <!-- Junction dot -->
  <circle cx="280" cy="130" r="4" fill="#000"/>
  <!-- Wire to Vout -->
  <line x1="280" y1="130" x2="380" y2="130" 
        stroke="#000" stroke-width="2"/>
  <circle cx="380" cy="130" r="4" fill="#000"/>
  <text x="395" y="134" font-family="Arial" font-size="12">Vout</text>
  <!-- Wire from junction down to C1 -->
  <line x1="280" y1="130" x2="280" y2="190" 
        stroke="#000" stroke-width="2"/>
  <!-- C1 capacitor plates -->
  <line x1="255" y1="190" x2="305" y2="190" 
        stroke="#000" stroke-width="2.5"/>
  <line x1="255" y1="205" x2="305" y2="205" 
        stroke="#000" stroke-width="2.5"/>
  <!-- C1 label to the right -->
  <text x="315" y="195" font-family="Arial" font-size="12" 
        font-weight="bold">C1</text>
  <text x="315" y="209" font-family="Arial" font-size="11">100nF</text>
  <!-- Wire from C1 down to GND -->
  <line x1="280" y1="205" x2="280" y2="260" 
        stroke="#000" stroke-width="2"/>
  <!-- GND symbol -->
  <line x1="255" y1="260" x2="305" y2="260" 
        stroke="#000" stroke-width="2"/>
  <line x1="263" y1="268" x2="297" y2="268" 
        stroke="#000" stroke-width="2"/>
  <line x1="271" y1="276" x2="289" y2="276" 
        stroke="#000" stroke-width="2"/>
</svg>

Draw circuits at appropriate scale — 
use the full 500x400 canvas, don't crowd everything in a corner.
For complex circuits use width="700" height="500".

--- "explanation" ---
4-5 sentence plain English explanation of how the circuit works, written entirely in {language} for a first-year EC student.

--- "spice_netlist" structure ---
A complete, ready-to-run ngspice netlist string.

Rules:
1. First line is always a comment: * CircuitName
2. Every component uses standard SPICE syntax:
   Resistors:   R1 node1 node2 10k
   Capacitors:  C1 node1 node2 100n  IC=0
   Inductors:   L1 node1 node2 1u    IC=0
   Diodes:      D1 anode cathode 1N4148
   NPN BJT:     Q1 collector base emitter NPN_MODEL
   MOSFET:      M1 drain gate source bulk NMOS_MODEL
   Op-amp:      Use subcircuit or behavioral model
   LED:         D1 anode cathode LED_MODEL
   Voltage src: Vin input 0 DC 5 AC 1

3. Node names: use short lowercase, no spaces
   VCC → vcc, GND → 0 (SPICE ground is always node 0)
   Component junctions: net1, net2, base, collector etc.

4. Always include these model definitions/subcircuits:
   .model 1N4148 D(Is=2.52n Rs=0.568 N=1.752)
   .model NPN_2N2222 NPN(Bf=200 Vaf=100 Is=14.34f)
   .model NPN_BC547  NPN(Bf=220 Vaf=80  Is=1e-14)
   .model PNP_BC557  PNP(Bf=200 Vaf=80  Is=1e-14)
   .model LED D(Is=1e-20 N=1.8 Rs=5)
   .subckt opamp_ideal vp vn vcc vee out
   E1 out 0 vp vn 1e5
   .ends opamp_ideal
   Only include models/subcircuits actually used in the circuit.
   For op-amps, instantiate them as: X1 vp vn vcc vee out opamp_ideal

5. Always include a voltage source:
   Vcc vcc 0 DC 5
   (use appropriate voltage for the circuit)

6. Include ALL THREE analysis types:
   .op
   .ac dec 100 1Hz 100MEGHz
   .tran {step} {stop} 0 {step}
   Choose step and stop based on circuit:
     RC filter 1kHz:  .tran 0.01m 5m
     555 timer 2Hz:   .tran 1m 2000m
     Oscillator 10MHz: .tran 1n 500n
     General:         .tran 1u 10m

7. Add .probe or .save for important nodes:
   .save v(vout) v(vcc) i(R1)

8. End with .end

Example for RC Low-Pass Filter:
* RC Low-Pass Filter 1kHz
Vin input 0 DC 0 AC 1 SIN(0 5 1000)
R1 input vout 1.59k
C1 vout 0 100n IC=0
Vcc vcc 0 DC 5
.op
.ac dec 100 1Hz 1MEGHz
.tran 0.01m 5m
.save v(vout) v(input)
.end

Return ONLY the JSON. No markdown. No backticks.
The spice_netlist value must be a single string 
with \n for newlines.

=== VERY IMPORTANT RULES ===
- Return ONLY the JSON object
- No markdown, no backticks, no text outside JSON
- The SVG must be valid XML — escape & as &amp; in text
- Every component in "circuit" must appear in "schematic_svg"
- Wires must NEVER pass through component bodies
- Labels must NEVER overlap wires or other labels
- For Op-Amps:
    - Draw a perfect triangle: <polygon points="x1,y1 x2,y2 x3,y3" fill="none" stroke="black" />
    - Label inputs explicitly with "+" and "-" next to pins
    - Feedback resistors (R_feedback) MUST be drawn as a clean right-angle path above the op-amp
    - Feedback wires MUST NOT overlap the main triangle body
- The spice_netlist MUST end exactly with .end and have NO text after it!
- If input is not circuit-related return: {"error": "Not a circuit"}
"""

MAX_RETRIES = 2
RETRY_DELAY = 5

# In-memory cache: { prompt_hash: (timestamp, result) }
_cache = {}
CACHE_TTL = 3600  # 1 hour

MODEL_CHAIN = [
    'gemini-flash-latest',
    'gemini-2.5-flash',
    'gemini-2.5-flash-lite',
]


def _call_gemini(prompt: str, model_index: int = 0, attempt: int = 0):
    """Call Gemini API with model fallback chain and retry logic."""
    model_name = MODEL_CHAIN[model_index]
    model = genai.GenerativeModel(model_name)
    try:
        print(f"[CircuitGPT] Trying {model_name}...")
        response = model.generate_content(prompt)
        print(f"[CircuitGPT] Success with {model_name}")
        return response
    except Exception as e:
        error_str = str(e).lower()
        is_rate_limit = ('quota' in error_str or 'rate' in error_str or
                         'resource' in error_str or '429' in error_str)

        if is_rate_limit:
            if model_index + 1 < len(MODEL_CHAIN):
                print(f"[CircuitGPT] {model_name} rate-limited, trying next model...")
                return _call_gemini(prompt, model_index + 1, attempt)

            if attempt < MAX_RETRIES:
                delay = RETRY_DELAY * (2 ** attempt)
                print(f"[CircuitGPT] All models rate-limited, waiting {delay}s (attempt {attempt + 1}/{MAX_RETRIES})...")
                time.sleep(delay)
                return _call_gemini(prompt, 0, attempt + 1)

        raise


def _clean_json(text: str) -> str:
    """Strip markdown code fences from Gemini response."""
    text = text.strip()
    text = re.sub(r'^```json\s*', '', text)
    text = re.sub(r'\s*```$', '', text)
    text = re.sub(r'^```\s*', '', text)
    return text.strip()


def _cache_key(description: str) -> str:
    return hashlib.md5(description.strip().lower().encode()).hexdigest()


def generate_circuit(description: str, language: str = "English") -> dict:
    """Generate circuit JSON from Gemini.
    Results are cached for CACHE_TTL seconds."""
    key = _cache_key(description + language)

    # Return cached result if fresh
    if key in _cache:
        ts, cached_result = _cache[key]
        if time.time() - ts < CACHE_TTL:
            print(f"[CACHE HIT] Returning cached result for: {description[:40]}")
            return cached_result

    prompt = f"""
Design this circuit: {description}

IMPORTANT: The explanation field must be written entirely in {language}. JSON keys stay in English.
"""
    system_prompt_formatted = SYSTEM_PROMPT.replace("{language}", language)

    response = _call_gemini(system_prompt_formatted + "\n\n" + prompt)
    raw = _clean_json(response.text)
    try:
        result = json.loads(raw)
        if "error" not in result:
            _cache[key] = (time.time(), result)
            print(f"[API CALL] Generated and cached: {description[:40]}")
        return result
    except json.JSONDecodeError:
        # Retry with stricter instruction
        response2 = _call_gemini(
            "Return ONLY a JSON object, no markdown:\n" +
            system_prompt_formatted + "\n\n" + prompt
        )
        raw2 = _clean_json(response2.text)
        result2 = json.loads(raw2)
        if "error" not in result2:
            _cache[key] = (time.time(), result2)
        return result2
