CirqAi-backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
python-multipart
openai-whisper
ffmpeg-python
numpy
